+++ overlay/overlay_text.py
@@ -1,9 +1,6 @@
 from __future__ import annotations
 
-from .config import OverlayConfig
-from .config import default_overlay_config
-from .config import normalize_overlay_config
-
+from .config import OverlayConfig, default_overlay_config, normalize_overlay_config
 
 SAMPLE_OVERLAY_VALUES = {
     "present_fps": "60",
@@ -131,7 +128,7 @@ def _ms_number(value: object) -> int | None:
     if text.lower().endswith(" ms"):
         text = text[:-3].strip()
     try:
-        return int(round(float(text)))
+        return round(float(text))
     except ValueError:
         return None
 
@@ -182,7 +179,7 @@ def _signed_value(value: object) -> str:
             text = text[: -len(suffix)].strip()
             break
     try:
-        number = int(round(float(text)))
+        number = round(float(text))
     except ValueError:
         return text
     return f"{number:+d}"
