+++ profiles/uv/runtime_auto_uv_profile.py
@@ -14,8 +14,8 @@ from auto_uv.gpu.memory_clock_offset_user_option import (
 from common.penguin_burner_errors import NvmlError
 from integrations.afterburner.policy import MAX_AFTERBURNER_MEM_OFFSET_MHZ
 
-from .profile_tiers import profile_tier_summary_fields
 from .profile_store import STOCK_PROFILE_SELECTOR, resolve_auto_uv_profile
+from .profile_tiers import profile_tier_summary_fields
 
 
 def load_auto_uv_final_curve(profile_selector="", *, allow_unverified: bool = False):
@@ -135,7 +135,7 @@ def profile_memory_offset_mhz(
         if value in (None, ""):
             continue
         try:
-            offset = max(0, int(round(float(value))))
+            offset = max(0, round(float(value)))
         except (TypeError, ValueError):
             return None
         if limit_mhz is not None:
@@ -149,7 +149,7 @@ def profile_power_limit_w(payload):
     if value in (None, ""):
         return None
     try:
-        power_limit = int(round(float(value)))
+        power_limit = round(float(value))
     except (TypeError, ValueError):
         return None
     return power_limit if power_limit > 0 else None
