+++ stability/q2rtx/resolution.py
@@ -38,13 +38,13 @@ def resolve_q2rtx_render_resolution(
     if width is not None:
         return Q2RTXResolutionChoice(
             width=width,
-            height=max(1, int(round(float(width) * 9.0 / 16.0))),
+            height=max(1, round(float(width) * 9.0 / 16.0)),
             reason="manual-width-16:9",
             auto_selected=False,
         )
     if height is not None:
         return Q2RTXResolutionChoice(
-            width=max(1, int(round(float(height) * 16.0 / 9.0))),
+            width=max(1, round(float(height) * 16.0 / 9.0)),
             height=height,
             reason="manual-height-16:9",
             auto_selected=False,
@@ -52,7 +52,7 @@ def resolve_q2rtx_render_resolution(
 
     try:
         memory_info = DaemonGpuClient(int(gpu_index)).capabilities().memory
-    except Exception:
+    except Exception:  # noqa: BLE001
         memory_info = None
     total_bytes = int(memory_info.total_bytes) if memory_info is not None else None
     if total_bytes is not None and total_bytes <= AUTO_RESOLUTION_MAX_1440P_BYTES:
